Move modules and fix FakeStorage
